refactor: move everything inside preview
