+++ src/components/preview/mail/mail-preview-content.tsx
@@ -9,13 +9,13 @@ import React, { useCallback, useMemo, useState } from 'react';
 import { Collapse, Container, Padding, Row } from '@zextras/carbonio-design-system';
 import { useIntegratedComponent, useUserSettings } from '@zextras/carbonio-shell-ui';
 
-import { AttachmentsBlock } from './attachments-block';
+import { AttachmentsBlock } from './attachments-block/attachments-block';
+import ReadReceiptModal from './read-receipt-modal';
 import { MailMessageRenderer } from '../../../commons/mail-message-renderer/mail-message-renderer';
 import { isFocusModeMailView } from '../../../helpers/external-tabs';
 import SharedInviteReply from '../../../integrations/shared-invite-reply';
 import { msgActionEmailStoreAction } from '../../../store/emails/actions/msg-action-action';
 import type { IncompleteMessage, MailMessage } from '../../../types';
-import ReadReceiptModal from '../../../views/app/preview/preview/read-receipt-modal';
 
 type MailPreviewContentProps = {
 	message: MailMessage | IncompleteMessage;
