+++ src/components/preview/mail/tests/mail-preview.test.tsx
@@ -9,11 +9,11 @@ import React from 'react';
 import { screen, within } from '@testing-library/react';
 import { FOLDERS } from '@zextras/carbonio-ui-commons';
 
-import MailPreview, { MailPreviewProps } from '../../../../../components/preview/mail/mail-preview';
+import { generateMessage } from '../../../../__test__/generators/generateMessage';
+import { getMsgSoapApi } from '../../../../api/get-msg-soap-api';
+import { normalizeMailMessageFromSoap } from '../../../../normalizations/normalize-message';
+import MailPreview, { MailPreviewProps } from '../mail-preview';
 import { setupTest } from '@test-setup';
-import { generateMessage } from '__test__/generators/generateMessage';
-import { getMsgSoapApi } from 'api/get-msg-soap-api';
-import { normalizeMailMessageFromSoap } from 'normalizations/normalize-message';
 
 /**
  * Test the Mail Preview component in different scenarios
