+++ src/helpers/external-tabs.ts
@@ -5,7 +5,7 @@
  */
 import { IS_FOCUS_MODE } from '@zextras/carbonio-shell-ui';
 
-import { getLocationOrigin } from '../views/app/preview/preview/utils/index';
+import { getLocationOrigin } from '../components/preview/mail/utils/index';
 import { EML_ROUTE, FOCUS_MODE_MAIL_VIEW_ROUTE, FOCUS_MODE_ROUTE } from 'constants/index';
 
 export const isFocusModeMailView = (): boolean => IS_FOCUS_MODE;
